rendering game_objects
changed the submit method to accept a game_object argument instead of the model argument. 

Then depending on whether the object has meshes or a vertex_array, the appropriate submit method is called. 

Added the transform method to the game_object to apply the stored rotation, position+offset and scale to the object.

Added texture bind method to game_object. Binds all textures if the object has any.

Changed the textured shader in main layer to include the normal in the layout.

Commented out the triangle/square/cube and multiple cube rendering.

Renders a single cube using the cube object class with a face texture on